Generic Logstash integration test framework
Can be hooked up to jenkins so it can be run every time a PR is submitted to change logstash configs.
Forward thinking towards Logstash 6 with 1 input mapped to 1 set of filters mapped to 1 output.

Tests are written by writing an elasticsearch query
Must provide an input (sample log to be ingested) and an output (truncated result from Elasticsearch).